refactor traversal into own directory
